+++ src/roq/deribit/multicast.cpp
@@ -2,11 +2,17 @@
 
 #include "roq/deribit/multicast.hpp"
 
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <sys/socket.h>
+
 #include "roq/core/metrics/factory.hpp"
 
 #include "roq/deribit/flags/common.hpp"
 #include "roq/deribit/flags/config.hpp"
 #include "roq/deribit/flags/multicast.hpp"
+
+#include "roq/deribit/sbe/utils.hpp"
 
 using namespace std::literals;
 
@@ -28,6 +34,14 @@ struct create_metrics final : public core::metrics::Factory {
 
 auto create_connection(auto &handler, auto &context, auto port) {
   core::net::UdpConnection connection(handler, context, port);
+  std::string local_interface{flags::Multicast::local_interface()};
+  struct in_addr local = {};
+  local.s_addr = inet_addr(local_interface.c_str());
+  for (auto &multicast_address : flags::Multicast::multicast_address()) {
+    struct in_addr multicast = {};
+    multicast.s_addr = inet_addr(multicast_address.c_str());
+    connection.add_membership(core::NetworkAddress{multicast, 0}, core::NetworkAddress{local, 0});
+  }
   return connection;
 }
 }  // namespace
@@ -56,6 +70,43 @@ void Multicast::operator()(const Event<Timer> &) {
 }
 
 void Multicast::operator()(const core::net::UdpConnection::Read &read) {
+  log::info<5>("received {} byte(s)"sv, std::size(read.buffer));
+  // XXX parse
+}
+
+void Multicast::operator()(
+    uint16_t channel_id, uint32_t sequence_number, deribit_multicast::Instrument &instrument) {
+  log::info<5>(
+      "channel_id={}, sequence_number={}, instrument={}"sv,
+      channel_id,
+      sequence_number,
+      instrument);
+}
+
+void Multicast::operator()(
+    uint16_t channel_id, uint32_t sequence_number, deribit_multicast::Book &book) {
+  log::info<5>("channel_id={}, sequence_number={}, book={}"sv, channel_id, sequence_number, book);
+}
+
+void Multicast::operator()(
+    uint16_t channel_id, uint32_t sequence_number, deribit_multicast::Quote &quote) {
+  log::info<5>("channel_id={}, sequence_number={}, quote={}"sv, channel_id, sequence_number, quote);
+}
+
+void Multicast::operator()(
+    uint16_t channel_id, uint32_t sequence_number, deribit_multicast::Trades &trades) {
+  /*
+  log::info<5>(
+      "channel_id={}, sequence_number={}, trades={}"sv, channel_id, sequence_number, trades);
+  */
+}
+
+void Multicast::operator()(
+    uint16_t channel_id, uint32_t sequence_number, deribit_multicast::Snapshot &snapshot) {
+  /*
+  log::info<5>(
+      "channel_id={}, sequence_number={}, snapshot={}"sv, channel_id, sequence_number, snapshot);
+  */
 }
 
 void Multicast::operator()(metrics::Writer &writer) {